requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
redis==3.5.3
six==1.16.0
requests==2.28.1
XlsxWriter==3.0.2
django-tenants==3.4.2
google-auth==2.9.1
oauth2client==4.1.3
sendgrid==6.9.7
python-dateutil==2.8.2
python-http-client==3.3.7
facebook-business==12.0.1
psycopg2-binary==2.9.4
selenium==4.2.0
webdriver-manager==3.7.0
boto3==1.26.26
Pillow==9.3.0
# get-video-properties==0.1.1
sentry-sdk==1.12.1
celery==5.2.7
flower
stripe==5.2.0
pdfkit==1.0.0
gunicorn==20.1.0
validate-email==1.3
hachoir==3.2.0
